// src/app/Models/_models/FlowItem.ts
export interface FlowItem {
    // id?: number,
    // eventtype: string,
    // display_name: string,
    // parent?: number,
    // root?: number,
    // level: number,
    // action: string,
    // intent: string,
    // slot:string,
    // response: string,
    id?: number;
    event_type: string;
    display_name: string;
    parent?: number;
    root?: number;
    level: number;
    flow: number;
    type_id: number;
    color: string;

}